Move ExtendedHandshake to separate file instead of putting it in Handshake.cs in Client
